fix project accessible check. Organized constants and internal classes
